% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS.fd.R
\name{ICS}
\alias{ICS}
\alias{ICS.default}
\title{Invariant Coordinate Selection (generic)}
\usage{
ICS(x, ...)

\method{ICS}{default}(x, ...)
}
\arguments{
\item{x}{The input data. The default method forwards to \code{\link[ICS:ICS-S3]{ICS::ICS()}}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class \code{ICS}.
}
\description{
Generic dispatching on the class of \code{x}. The default method is
\code{\link[ICS:ICS-S3]{ICS::ICS()}}; methods are provided for \code{fd} and \code{acomp} objects.
}
\seealso{
\code{\link[ICS:ICS-S3]{ICS::ICS()}}, \code{\link[=ICS.fd]{ICS.fd()}}, \code{\link[=ICS.acomp]{ICS.acomp()}}
}
