% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_outlier.fd.R
\name{ICS_outlier}
\alias{ICS_outlier}
\alias{ICS_outlier.default}
\title{ICS outlier detection (generic)}
\usage{
ICS_outlier(x, ...)

\method{ICS_outlier}{default}(x, ...)
}
\arguments{
\item{x}{The input data.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object describing detected outliers (class depends on method).
}
\description{
Generic dispatching on the class of \code{x}. The default method is
\code{\link[ICSOutlier:ICS_outlier]{ICSOutlier::ICS_outlier()}}; a method is provided for \code{fd} objects.
}
\seealso{
\code{\link[ICSOutlier:ICS_outlier]{ICSOutlier::ICS_outlier()}}, \code{\link[=ICS_outlier.fd]{ICS_outlier.fd()}}
}
