% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_outlier.fd.R
\name{mdist_simu_test}
\alias{mdist_simu_test}
\title{Memoised version of \code{\link[ICSOutlier:dist_simu_test]{ICSOutlier::dist_simu_test()}}}
\usage{
mdist_simu_test(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ICSOutlier:dist_simu_test]{ICSOutlier::dist_simu_test()}}.}
}
\value{
The value returned by \code{\link[ICSOutlier:dist_simu_test]{ICSOutlier::dist_simu_test()}}.
}
\description{
A thin wrapper around \code{\link[ICSOutlier:dist_simu_test]{ICSOutlier::dist_simu_test()}} that memoises results
so repeated calls with identical arguments are returned from cache.
}
\seealso{
\code{\link[ICSOutlier:dist_simu_test]{ICSOutlier::dist_simu_test()}}
}
