% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vietnam_regions}
\alias{vietnam_regions}
\title{Vietnam administrative regions}
\format{
A data frame with 6 rows and 2 columns:
\describe{
\item{code}{Region code.}
\item{name}{Region name.}
}
}
\usage{
vietnam_regions
}
\description{
Lookup table mapping administrative region codes to names.
}
\keyword{datasets}
