% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vietnam_temperature}
\alias{vietnam_temperature}
\title{Vietnam daily temperature records}
\format{
A tibble with 1890 rows and 5 columns:
\describe{
\item{year}{Year of observation.}
\item{region}{Administrative region code.}
\item{province}{Province name.}
\item{t_max}{List-column of daily maximum temperatures.}
\item{t_min}{List-column of daily minimum temperatures.}
}
}
\usage{
vietnam_temperature
}
\description{
Annual records of daily minimum and maximum temperatures by province.
}
\keyword{datasets}
