% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vietnam_temperature_dd}
\alias{vietnam_temperature_dd}
\title{Vietnam temperatures as distributional data}
\format{
A rowwise tibble with 1890 rows and 5 columns:
\describe{
\item{year}{Year of observation.}
\item{region}{Administrative region code.}
\item{province}{Province name.}
\item{t_max}{\code{ddl} list-column of daily maximum temperatures.}
\item{t_min}{\code{ddl} list-column of daily minimum temperatures.}
}
}
\usage{
vietnam_temperature_dd
}
\description{
Same data as \link{vietnam_temperature} with \code{t_max} and \code{t_min} stored as
distributional (\code{ddl}) list-columns.
}
\keyword{datasets}
